moves apply_custom_transformations to process.py
